Cursor Scrubber/requirements.txt
PyMuPDF>=1.23.0
opencv-python-headless>=4.8.0
pytesseract>=0.3.10
Pillow>=10.0.0
streamlit>=1.28.0
numpy>=1.24.0
regex==2023.8.8
python-docx==0.8.11
reportlab==4.0.4
# Enhanced dependencies for better accuracy
pdfplumber==0.10.3
ultralytics==8.0.196
spacy==3.7.2
paddlepaddle==2.5.2
paddleocr==2.7.0.3
scikit-learn==1.3.0
matplotlib==3.7.2
seaborn==0.12.2
# Built-in modules (no need to install):
# sqlite3, zipfile, io, hashlib, uuid, datetime, typing 